fix(misra): 10.4 批1 pass-2 — 工具升级（enum 常量/case 标签/char 字面量/宏定位/多分支宏体），266→97

+++ src/asw/diagnostic_manager/src/Swc_DiagnosticManager.c
@@ -523,7 +523,7 @@ void Swc_DiagnosticManager_ProcessRequest(void)
 
     if (!swcDiagManager.hasPendingRequest) {
         /* Check for new request via RTE */
-        if ((uint8_t)((uint8_t)(Rte_Read_DiagnosticRequest(&request))) == RTE_E_OK) {
+        if ((uint8_t)((uint8_t)((uint8_t)(Rte_Read_DiagnosticRequest(&(uint32_t)(request))))) == RTE_E_OK) {
             swcDiagManager.pendingRequest = request;
             swcDiagManager.hasPendingRequest = TRUE;
         } else {
